OK now

+++ core/src/main/java/com/ninetwozero/bf4intel/json/soldieroverview/SkillOverview.java
@@ -54,7 +54,7 @@ public class SkillOverview {
     }
 
     public double getKillsPerMinute() {
-        return ((double) killCount) / (timePlayed / 60);
+        return ((double) killCount) / ((double)timePlayed / 60);
     }
 
     public Map<Integer, Integer> getServiceStars() {
